fix: Support different version builds with Docker
Currently, when attempting to switch back and forth between 9.x and
10.0.x, many Docker volumes and services are shared across the builds.
This makes it impossible to reproducibly test the differences between
these versions of OnTrack. Many of the changes made here are naming
convention changes to assist Docker and vscode to reliably create
distinct containers for both 9.x and 10.0.x. This change is intended
to assist developers in comparing and testing the differences between
these divergent development branches.